requirements.txt
asgiref==3.7.2
Django==4.2.2
grpcio==1.56.0
grpcio-tools==1.56.0
protobuf==4.23.3
sqlparse==0.4.4
typing_extensions==4.6.3
